fix(bayer): cast the demosaiced image back to the input dtype
colour_demosaicing runs in colour-science's default float dtype, so debayering a float32 mosaic returned float64 — against what the docstring promised, and enough to make every converted CFA frame a BITPIX -64 file, twice the size it should be. Casting before the clip also spares the second float64 array.

+++ umbra/common/bayer.py
@@ -55,6 +55,7 @@ def debayer(img: np.ndarray, pattern: str, algorithm: str = "bilinear") -> np.nd
     except KeyError as exc:
         raise ValueError(f"Unsupported demosaicing algorithm {algorithm}.") from exc
     print("Debayering...", end="", flush=True)
-    rgb = demosaic(img, pattern)
+    # colour_demosaicing works in colour-science's default float dtype (float64).
+    rgb = demosaic(img, pattern).astype(img.dtype, copy=False)
     print("Done.")
     return np.clip(rgb, 0, 1)
